REFACTORY Grid Node Application (#264)
