+++ app/websocket/app/templates/index.html
@@ -1,0 +1,8 @@
+<html>
+    <head>
+        <title>Grid Node</title>
+    </head>
+    <body>
+        <h1>Websocket Grid Node!</h1>
+    </body>
+</html>
